Refactor the generic query function; organize functions to appropriate files
