+++ client.go
@@ -23,7 +23,6 @@ use case. Pull requests welcome :)
 package quickbooks
 
 import (
-	"bytes"
 	"encoding/json"
 	"errors"
 	"io/ioutil"
@@ -79,311 +78,22 @@ func (c *Client) FetchCompanyInfo() (*CompanyInfo, error) {
 	return &r.CompanyInfo, err
 }
 
-// FetchCustomers gets the full list of Customers in the QuickBooks account.
-func (c *Client) FetchCustomers() ([]Customer, error) {
+// query makes the specified QBO `query` and unmarshals the result into `out`
+func (c *Client) query(query string, out interface{}) error {
 	var u, err = url.Parse(string(c.Endpoint))
 	if err != nil {
-		return nil, err
-	}
-	u.Path = "/v3/company/" + c.RealmID + "/query"
-
-	// See how many customers there are.
-	var v = url.Values{}
-	v.Add("query", "SELECT COUNT(*) FROM Customer")
-	u.RawQuery = v.Encode()
-	var req *http.Request
-	req, err = http.NewRequest("GET", u.String(), nil)
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Add("Accept", "application/json")
-	var res *http.Response
-	res, err = c.Client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer res.Body.Close()
-
-	// TODO This could be better...
-	if res.StatusCode != http.StatusOK {
-		var msg []byte
-		msg, err = ioutil.ReadAll(res.Body)
-		return nil, errors.New(strconv.Itoa(res.StatusCode) + " " + string(msg))
-	}
-
-	var r struct {
-		QueryResponse struct {
-			TotalCount int
-		}
-	}
-	err = json.NewDecoder(res.Body).Decode(&r)
-	if err != nil {
-		return nil, err
-	}
-
-	if r.QueryResponse.TotalCount == 0 {
-		return make([]Customer, 0), nil
-	}
-
-	var customers = make([]Customer, 0, r.QueryResponse.TotalCount)
-	for i := 0; i < r.QueryResponse.TotalCount; i += queryPageSize {
-		var page, err = c.fetchCustomerPage(i + 1)
-		if err != nil {
-			return nil, err
-		}
-		customers = append(customers, page...)
-	}
-	return customers, nil
-}
-
-// Fetch one page of results, because we can't get them all in one query.
-func (c *Client) fetchCustomerPage(startpos int) ([]Customer, error) {
-	var u, err = url.Parse(string(c.Endpoint))
-	if err != nil {
-		return nil, err
+		return err
 	}
 	u.Path = "/v3/company/" + c.RealmID + "/query"
 
 	var v = url.Values{}
-	v.Add("query", "SELECT * FROM Customer ORDERBY Id STARTPOSITION "+
-		strconv.Itoa(startpos)+" MAXRESULTS "+strconv.Itoa(queryPageSize))
+	v.Add("query", query)
 	u.RawQuery = v.Encode()
 	var req *http.Request
 	req, err = http.NewRequest("GET", u.String(), nil)
 	if err != nil {
-		return nil, err
-	}
-	req.Header.Add("Accept", "application/json")
-	var res *http.Response
-	res, err = c.Client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer res.Body.Close()
-
-	// TODO This could be better...
-	if res.StatusCode != http.StatusOK {
-		var msg []byte
-		msg, err = ioutil.ReadAll(res.Body)
-		return nil, errors.New(strconv.Itoa(res.StatusCode) + " " + string(msg))
-	}
-
-	var r struct {
-		QueryResponse struct {
-			Customer      []Customer
-			StartPosition int
-			MaxResults    int
-		}
-	}
-	err = json.NewDecoder(res.Body).Decode(&r)
-	if err != nil {
-		return nil, err
-	}
-	// Make sure we don't return nil if there are no customers.
-	if r.QueryResponse.Customer == nil {
-		r.QueryResponse.Customer = make([]Customer, 0)
-	}
-	return r.QueryResponse.Customer, nil
-}
-
-// CreateCustomer creates the given Customer on the QuickBooks server,
-// returning the resulting Customer object.
-func (c *Client) CreateCustomer(customer *Customer) (*Customer, error) {
-	var u, err = url.Parse(string(c.Endpoint))
-	if err != nil {
-		return nil, err
-	}
-	u.Path = "/v3/company/" + c.RealmID + "/customer"
-	var j []byte
-	j, err = json.Marshal(customer)
-	if err != nil {
-		return nil, err
-	}
-	var req *http.Request
-	req, err = http.NewRequest("POST", u.String(), bytes.NewBuffer(j))
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Add("Content-Type", "application/json")
-	req.Header.Add("Accept", "application/json")
-	var res *http.Response
-	res, err = c.Client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer res.Body.Close()
-
-	// TODO This could be better...
-	if res.StatusCode != http.StatusOK {
-		var msg []byte
-		msg, err = ioutil.ReadAll(res.Body)
-		return nil, errors.New(strconv.Itoa(res.StatusCode) + " " + string(msg))
-	}
-
-	var r struct {
-		Customer Customer
-		Time     time.Time
-	}
-	err = json.NewDecoder(res.Body).Decode(&r)
-	return &r.Customer, err
-}
-
-// FetchItems returns the list of Items in the QuickBooks account. These are
-// basically product types, and you need them to create invoices.
-func (c *Client) FetchItems() ([]Item, error) {
-	var u, err = url.Parse(string(c.Endpoint))
-	if err != nil {
-		return nil, err
-	}
-	u.Path = "/v3/company/" + c.RealmID + "/query"
-
-	var v = url.Values{}
-	v.Add("query", "SELECT * FROM Item MAXRESULTS "+strconv.Itoa(queryPageSize))
-	u.RawQuery = v.Encode()
-	var req *http.Request
-	req, err = http.NewRequest("GET", u.String(), nil)
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Add("Accept", "application/json")
-	var res *http.Response
-	res, err = c.Client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer res.Body.Close()
-
-	// TODO This could be better...
-	if res.StatusCode != http.StatusOK {
-		var msg []byte
-		msg, err = ioutil.ReadAll(res.Body)
-		return nil, errors.New(strconv.Itoa(res.StatusCode) + " " + string(msg))
-	}
-
-	var r struct {
-		QueryResponse struct {
-			Item          []Item
-			StartPosition int
-			MaxResults    int
-		}
-	}
-	err = json.NewDecoder(res.Body).Decode(&r)
-	if err != nil {
-		return nil, err
-	}
-	// Make sure we don't return nil if there are no items.
-	if r.QueryResponse.Item == nil {
-		r.QueryResponse.Item = make([]Item, 0)
-	}
-	return r.QueryResponse.Item, nil
-}
-
-// FetchItem returns just one particular Item from QuickBooks, by ID.
-func (c *Client) FetchItem(id string) (*Item, error) {
-	var u, err = url.Parse(string(c.Endpoint))
-	if err != nil {
-		return nil, err
-	}
-	u.Path = "/v3/company/" + c.RealmID + "/item/" + id
-
-	var req *http.Request
-	req, err = http.NewRequest("GET", u.String(), nil)
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Add("Accept", "application/json")
-	var res *http.Response
-	res, err = c.Client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer res.Body.Close()
-
-	// TODO This could be better...
-	if res.StatusCode != http.StatusOK {
-		var msg []byte
-		msg, err = ioutil.ReadAll(res.Body)
-		return nil, errors.New(strconv.Itoa(res.StatusCode) + " " + string(msg))
-	}
-
-	var r struct {
-		Item Item
-		Time time.Time
-	}
-	err = json.NewDecoder(res.Body).Decode(&r)
-	if err != nil {
-		return nil, err
-	}
-	return &r.Item, nil
-}
-
-// CreateInvoice creates the given Invoice on the QuickBooks server, returning
-// the resulting Invoice object.
-func (c *Client) CreateInvoice(inv *Invoice) (*Invoice, error) {
-	var u, err = url.Parse(string(c.Endpoint))
-	if err != nil {
-		return nil, err
-	}
-	u.Path = "/v3/company/" + c.RealmID + "/invoice"
-	var j []byte
-	j, err = json.Marshal(inv)
-	if err != nil {
-		return nil, err
-	}
-	var req *http.Request
-	req, err = http.NewRequest("POST", u.String(), bytes.NewBuffer(j))
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Add("Content-Type", "application/json")
-	req.Header.Add("Accept", "application/json")
-	var res *http.Response
-	res, err = c.Client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer res.Body.Close()
-
-	// TODO This could be better...
-	if res.StatusCode != http.StatusOK {
-		var msg []byte
-		msg, err = ioutil.ReadAll(res.Body)
-		return nil, errors.New(strconv.Itoa(res.StatusCode) + " " + string(msg))
-	}
-
-	var r struct {
-		Invoice Invoice
-		Time    time.Time
-	}
-	err = json.NewDecoder(res.Body).Decode(&r)
-	return &r.Invoice, err
-}
-
-// DeleteInvoice deletes the given Invoice by ID and sync token from the
-// QuickBooks server.
-func (c *Client) DeleteInvoice(id, syncToken string) error {
-	var u, err = url.Parse(string(c.Endpoint))
-	if err != nil {
 		return err
 	}
-	u.Path = "/v3/company/" + c.RealmID + "/invoice"
-	var j []byte
-	j, err = json.Marshal(struct {
-		ID        string `json:"Id"`
-		SyncToken string
-	}{
-		ID:        id,
-		SyncToken: syncToken,
-	})
-	if err != nil {
-		return err
-	}
-	var req *http.Request
-	req, err = http.NewRequest("POST", u.String()+"?operation=delete", bytes.NewBuffer(j))
-	if err != nil {
-		return err
-	}
-	req.Header.Add("Content-Type", "application/json")
 	req.Header.Add("Accept", "application/json")
 	var res *http.Response
 	res, err = c.Client.Do(req)
@@ -391,37 +101,7 @@ func (c *Client) DeleteInvoice(id, syncToken string) error {
 		return err
 	}
 	defer res.Body.Close()
-	//var b, _ = ioutil.ReadAll(res.Body)
-	//log.Println(string(b))
 
-	// If the invoice was already deleted, QuickBooks returns 400 :(
-	// The response looks like this:
-	// {"Fault":{"Error":[{"Message":"Object Not Found","Detail":"Object Not Found : Something you're trying to use has been made inactive. Check the fields with accounts, customers, items, vendors or employees.","code":"610","element":""}],"type":"ValidationFault"},"time":"2018-03-20T20:15:59.571-07:00"}
-
-	// This is slightly horrifying and not documented in their API. When this
-	// happens we just return success; the goal of deleting it has been
-	// accomplished, just not by us.
-	if res.StatusCode == http.StatusBadRequest {
-		var r struct {
-			Fault struct {
-				Error []struct {
-					Message string
-					Detail  string
-					Code    string `json:"code"`
-					Element string `json:"element"`
-				}
-				Type string `json:"type"`
-			}
-			Time time.Time `json:"time"`
-		}
-		err = json.NewDecoder(res.Body).Decode(&r)
-		if err != nil {
-			return err
-		}
-		if r.Fault.Error[0].Message == "Object Not Found" {
-			return nil
-		}
-	}
 	// TODO This could be better...
 	if res.StatusCode != http.StatusOK {
 		var msg []byte
@@ -429,6 +109,5 @@ func (c *Client) DeleteInvoice(id, syncToken string) error {
 		return errors.New(strconv.Itoa(res.StatusCode) + " " + string(msg))
 	}
 
-	// TODO they send something back, but is it useful?
-	return nil
+	return json.NewDecoder(res.Body).Decode(out)
 }
