+++ item.go
@@ -3,6 +3,16 @@
 
 package quickbooks
 
+import (
+	"encoding/json"
+	"errors"
+	"io/ioutil"
+	"net/http"
+	"net/url"
+	"strconv"
+	"time"
+)
+
 // Item represents a QuickBooks Item object (a product type).
 type Item struct {
 	ID        string `json:"Id,omitempty"`
@@ -31,4 +41,65 @@ type Item struct {
 	QtyOnHand          float32       `json:",omitempty"`
 	SalesTaxCodeRef    ReferenceType `json:",omitempty"`
 	PurchaseTaxCodeRef ReferenceType `json:",omitempty"`
+}
+
+// FetchItems returns the list of Items in the QuickBooks account. These are
+// basically product types, and you need them to create invoices.
+func (c *Client) FetchItems() ([]Item, error) {
+	var r struct {
+		QueryResponse struct {
+			Item          []Item
+			StartPosition int
+			MaxResults    int
+		}
+	}
+	err := c.query("SELECT * FROM Item MAXRESULTS "+strconv.Itoa(queryPageSize), &r)
+	if err != nil {
+		return nil, err
+	}
+
+	// Make sure we don't return nil if there are no items.
+	if r.QueryResponse.Item == nil {
+		r.QueryResponse.Item = make([]Item, 0)
+	}
+	return r.QueryResponse.Item, nil
+}
+
+// FetchItem returns just one particular Item from QuickBooks, by ID.
+func (c *Client) FetchItem(id string) (*Item, error) {
+	var u, err = url.Parse(string(c.Endpoint))
+	if err != nil {
+		return nil, err
+	}
+	u.Path = "/v3/company/" + c.RealmID + "/item/" + id
+
+	var req *http.Request
+	req, err = http.NewRequest("GET", u.String(), nil)
+	if err != nil {
+		return nil, err
+	}
+	req.Header.Add("Accept", "application/json")
+	var res *http.Response
+	res, err = c.Client.Do(req)
+	if err != nil {
+		return nil, err
+	}
+	defer res.Body.Close()
+
+	// TODO This could be better...
+	if res.StatusCode != http.StatusOK {
+		var msg []byte
+		msg, err = ioutil.ReadAll(res.Body)
+		return nil, errors.New(strconv.Itoa(res.StatusCode) + " " + string(msg))
+	}
+
+	var r struct {
+		Item Item
+		Time time.Time
+	}
+	err = json.NewDecoder(res.Body).Decode(&r)
+	if err != nil {
+		return nil, err
+	}
+	return &r.Item, nil
 }
